Moved another public class to internal package
So that we reduce public API footprint
