+++ src/test/groovy/org/mockito/release/notes/format/DefaultFormatterTest.groovy
@@ -1,7 +1,7 @@
 package org.mockito.release.notes.format
 
 import org.mockito.release.notes.internal.DefaultReleaseNotesData
-import org.mockito.release.notes.improvements.DefaultImprovement
+import org.mockito.release.notes.internal.DefaultImprovement
 import org.mockito.release.notes.util.Predicate
 import org.mockito.release.notes.model.ContributionSet
 import org.mockito.release.notes.vcs.DefaultContributionSet
